add aggregation implementation + it-test

+++ modules/core/services/src/main/scala/org/corespring/services/item/ItemAggregationService.scala
@@ -6,7 +6,7 @@ import scala.concurrent.Future
 
 trait ItemAggregationService {
 
-  def contributorCounts(collectionIds: Seq[ObjectId]): Future[Map[String, Int]]
+  def contributorCounts(collectionIds: Seq[ObjectId]): Future[Map[String, Double]]
 
-  def taskInfoItemTypeCounts(collectionIds: Seq[ObjectId]): Future[Map[String, Int]]
+  def taskInfoItemTypeCounts(collectionIds: Seq[ObjectId]): Future[Map[String, Double]]
 }
